move header to model package
