+++ core/src/main/scala/zio/http/model/Response.scala
@@ -1,7 +1,5 @@
 package zio.http.model
 
-import zio.http.Header
-
 final case class Response[T](
   headers: List[Header],
   status: StatusCode,
